chore: Move `replay-block` back to stacks-inspect
